pyproject.toml
[project]
name = "miroeval"
version = "2.0.0"
description = "MiroEval: Benchmarking Multimodal Deep Research Agents in Process and Outcome"
license = "Apache-2.0"
requires-python = ">=3.11,<3.14"
dependencies = [
    # Core (used by quality + process evaluation)
    "openai>=1.0",
    "python-dotenv>=1.0",
    "pyyaml>=6.0",
    "tqdm>=4.60",
]

[project.optional-dependencies]
# Factual evaluation requires the full MiroFlow agent framework.
factual = ["miroflow>=1.6.0"]
# Install everything (for development or full evaluation).
all = ["miroeval[factual]"]

[project.scripts]
miroeval = "miroeval.cli:main"

[tool.uv.sources]
# During development, use the local miroflow package.
# For production, remove this and install miroflow from PyPI/git.
miroflow = { path = "factual_eval", editable = true }

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["miroeval"]
